Modifications to avoid the long chain of cases for Either results.
It uses the monad transforme ExcepT to use (ExceptT Error IO a) -- in the end only IO (Either Error a) wrapped...
Primitive functions where not altered, but modified with "ExceptT $ return $";
it is expected that a full change will change the type of primitive functions to be ExcepT, so this is not needed.

New chains of ExceptT computations allow for the detection of common prefixes in computation, so new intermediate functions may be created.